Created a lot of VM menu items.
Created a list of VMs. Show shutdown when it is shutdown.
When the focus is on a VM name, information is displayed next to it.
The information displayed is CPU usage, memory usage, disk usage, and network usage.
Currently supports width resizing, but does not support vertical resizing.

Multiple disks can be displayed, but NICs are not supported.